Move Aries Blueprint namespace handler into a separate akka-osgi-aries module
