+++ akka-osgi-aries/src/main/resources/akka/osgi/aries/blueprint/akka.xsd
