+++ akka-osgi-aries/src/test/resources/akka/osgi/aries/blueprint/config.xml
