+++ akka-osgi-aries/src/test/resources/akka/osgi/aries/blueprint/injection.xml
@@ -2,8 +2,8 @@
 <blueprint xmlns="http://www.osgi.org/xmlns/blueprint/v1.0.0"
            xmlns:akka="http://akka.io/xmlns/blueprint/v1.0.0">
 
-    <service interface="akka.osgi.test.ActorSystemAwareBean">
-        <bean class="akka.osgi.test.ActorSystemAwareBean">
+    <service interface="akka.osgi.aries.blueprint.ActorSystemAwareBean">
+        <bean class="akka.osgi.aries.blueprint.ActorSystemAwareBean">
             <argument>
                 <akka:actor-system name="simple" />
             </argument>
