+++ akka-osgi-aries/src/test/scala/akka/osgi/aries/blueprint/ActorSystemAwareBean.scala
@@ -1,4 +1,4 @@
-package akka.osgi.test
+package akka.osgi.aries.blueprint
 
 import akka.actor.ActorSystem
